% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGraph.R
\name{dGraph}
\alias{dGraph}
\title{Dynamic Graph}
\usage{
dGraph(seq, dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTGGG" or as a FASTA type}

\item{dim}{- number of dimensions: 2 (default) or 3}

\item{vec}{- list of unit vectors associated with nucleobases}

\item{type}{'s' - raw sequence, 'f' - fasta}
}
\value{
A list containing two matrices: 'coordinates' and 'graph'
}
\description{
Create representation
}
