% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drep.R
\name{drep}
\alias{drep}
\title{2D-DYnamic Representation of the DNA/RNA Sequences}
\usage{
drep(seq, type = "sequence", dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTTGAGGGAG"}

\item{type}{DNA or RNA}
}
\description{
2D-DYnamic Representation of the DNA/RNA Sequences
}
