% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drepMultiple.R
\name{drepMultiple}
\alias{drepMultiple}
\title{Dynamic Representation of Multiple Sequences}
\usage{
drepMultiple(seqs, type = 's', dim = 2)
}
\arguments{
\item{seqs}{sequences given as a vector of single strings, e.g. c('AGTGGG', 'TTGA') or as a FASTA type}

\item{type}{'s' - symbolic, 'wb' - grayscale, 'c' - colors}
}
\value{

}
\description{
Characterize multiple sequences at the same time
}
