% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2DGraph.R
\name{plot2DGraph}
\alias{plot2DGraph}
\title{Plot 2D-Dynamic Graph}
\usage{
plot2DGraph(seq, type = 's', dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTGGG" or as a FASTA type}

\item{type}{'s' - symbolic, 'wb' - grayscale, 'c' - colors}
}
\value{

}
\description{
Create plot
}
