% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGraph.R
\name{dGraph}
\alias{dGraph}
\title{Dynamic Graph}
\usage{
dGraph(seq, dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTGGG" or as a FASTA type}

\item{type}{'s' - raw sequence, 'f' - fasta}
}
\value{
A matrix of size Nx(dim+1) where N is the number of unique positions on the graph
}
\description{
Create representation
}
