% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRep.R
\name{dRep}
\alias{dRep}
\title{Dynamic Representation of the DNA/RNA Sequences}
\usage{
drep(seq, type = 's', dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTGGG" or as a FASTA type}

\item{type}{'s' - raw sequence, 'f' - fasta}
}
\value{
A matrix of size 2xN or 3xN, where N is the length of the sequence, giving position of every step of the walk.
}
\description{
Create representation
}
