% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRep.R
\name{dRep}
\alias{dRep}
\title{Representation of the DNA/RNA Sequences}
\usage{
dRep(seqs, dim = 2, genbank = FALSE, vec = list('A' = c(-1, 0, 1),
'C' = c(0, 1, 1), 'G' = c(1, 0, 1), 'T' = c(0, -1, 1)))
}
\arguments{
\item{seqs}{vector of sequences given as strings or GenBank accession numbers}

\item{dim}{number of dimensions: 2 (default) or 3}

\item{genbank}{FALSE (default) - sequence, TRUE - accession number from GenBank}

\item{vec}{list of unit vectors associated with nucleobases}
}
\value{
Dataframe with N row and 14 columns (dim = 2) or 26 columns (dim = 3), where each row represents one of N given sequences
}
\description{
dRep creates characterization of the 2D- or 3D-Dynamic Graph. The characterization can be used in an alignment-free analysis of the DNA/RNA sequences.
}
