% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRep.R
\name{dRep}
\alias{dRep}
\title{Dynamic Representation of the DNA/RNA Sequences}
\usage{
drep(seq, type = 's', dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTGGG" or as a FASTA file}

\item{dim}{number of dimensions: 2 (default) or 3}

\item{genbank}{FALSE (default) - raw sequence, TRUE - accession number from GenBank}

\item{vec}{list of unit vectors associated with nucleobases}
}
\value{
Dataframe with one row and 14 columns (dim = 2) or 26 columns (dim = 3) with descriptors
}
\description{
dRep reates characterization of the dynamic graph that is built by dGraph function. The characterization can be used in an alignment-free analysis of the DNA/RNA sequences.
}
